-- changed VON_KARMAK_STRAIN to NONLINEAR_STRAIN.     -- 3D structural element now excludes the nonlinear terms if strain in the property card is set to NONLINEAR_STRAIN. (Thank you, John Neiferd, for identifying this bug).

+++ src/elasticity/structural_transient_assembly.cpp
@@ -177,7 +177,7 @@ elem_sensitivity_calculations(const MAST::FunctionBase& f,
     dynamic_cast<MAST::StructuralElementBase&>(*_physics_elem);
     
     unsigned int
-    n       =  f_m.size();
+    n       =  (unsigned int)f_m.size();
     
     RealMatrixX
     dummy   =  RealMatrixX::Zero(n, n);
